Move host services to it's own package
